MWPW-147289: [Nala] Refine screenshot diff tool, clean take.js and update compare.js and add timestamp
