+++ tests/visual/caas/cards.test.js
@@ -1,7 +1,8 @@
 /* eslint-disable import/named */
 import { test } from '@playwright/test';
 import { features } from '../../../features/visual/caas/cards.spec.js';
-import { takeTwoAndCompare, writeResultsToFile } from '../../../libs/screenshot/take.js';
+import { takeTwo } from '../../../libs/screenshot/take.js';
+import { writeResultsToFile } from '../../../libs/screenshot/utils.js';
 
 const folderPath = 'screenshots/caas';
 const results = {};
@@ -12,7 +13,7 @@ test.describe('Milo Caas block visual comparison test suite', () => {
     // eslint-disable-next-line no-loop-func
     test(`${feature.name},${feature.tags}`, async ({ page, baseURL }, testInfo) => {
       const name = `${feature.name}-${testInfo.project.name}`;
-      const result = await takeTwoAndCompare(
+      const result = await takeTwo(
         page,
         baseURL + feature.stable,
         async () => { await page.waitForTimeout(3000); },
